Split diffing to separate module
